Update `change` event types to allow `ChangePayload` & documentation updates
The `change` function only accepted the `EventParameters` parameter type which meant that it couldn't be used to send a Change event as per the API documentation which doesn't require a severity to be passed. The `change` function now uses the `ChangeParameters` type for it's parameter and the tests have been updated/added accordingly. This will resolve #336.

Small changes made to documentation to reflect the above as well as giving an example of how to add headers to a request (related to #276). Resolved a few markdown linting issues too.

Signed-off-by: Aaron Parker <aaron@sourcecube.co.uk>